pyproject.toml
[project]
name = "ee-bench"
version = "0.1.0"
description = "Explore/Exploit Benchmark for LLMs"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "matplotlib>=3.8",
    "numpy>=1.26",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6.0",
]
analysis = [
    "matplotlib>=3.8",
    "pandas>=2.1",
    "scipy>=1.12",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
ee-bench = "ee_bench.cli:main"
ee-bench-analyze = "ee_bench.analysis:main"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
